refactor: put the command in same file as completion
